+++ CHANGELOG.md
@@ -5,6 +5,10 @@
  - Fix inability to switch any output except channel 0 from web interface ([#840](https://github.com/kizniche/mycodo/issues/840))
  - Minor fixes for PCF8574 Output
 
+### Features
+
+ - Add ability to select method in Input/Output/Function controller custom options
+
 
 ## 8.8.1 (2020-09-09)